+++ docker-compose.yml
@@ -89,6 +89,6 @@ services:
     container_name: migrate
     volumes:
       - ./migrate:/data
-    entrypoint: ["/bin/sh", "-c", "apk add --no-cache curl && chmod +x /data/* && /data/valheim.sh"]
+    entrypoint: ["/bin/sh", "-c", "/data/valheim.sh"]
     depends_on:
       - stoker
